refactor: move file loading utility to dom.js and remove unused Firebase storage and database modules
